requirements.txt
faster-whisper>=1.0.0
pyannote.audio>=3.1,<4
pydub>=0.25.1
click>=8.1.0
python-dotenv>=1.0.0
torch>=2.0,<2.6
torchaudio>=2.0,<2.6
huggingface_hub>=0.20,<0.26

# Optional: only needed if using --summary flag for automatic API summarization
# anthropic>=0.40.0
